Pull in new test result path for Xcode 10. (#13280)
Fixes #13034.

### Checklist
- [x] I've run `bundle exec rspec` from the root directory to see all new and existing tests pass
- [x] I've followed the _fastlane_ code style and run `bundle exec rubocop -a` to ensure the code style is valid
- [x] I've read the [Contribution Guidelines](https://github.com/fastlane/fastlane/blob/master/CONTRIBUTING.md)
- [x] I've updated the documentation if necessary.

### Motivation and Context
Xcode 10 changes where test reports are placed, and this breaks existing code that locates the test report plist files. This PR looks in both the old and new locations to determine where the test logs live.

### Description
The `test_summary_filenames` method now builds an array of files from both locations. This is purely additive, so the existing code should continue to work fine on Xcode 9.